+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 - pass module context (module descriptor) to builtin ``complete`` command via ``-m`` option
 - pass module context of current completion to ``COMP_HOOK``
 
+#### LSP
+
+- support ``textDocument/documentHighlight``
+
 #### Misc
 
 - add standalone syntax highlighter called ``dscolorize``